refactor(task)!: invert the entry split — node-free base, ./node interpreters
Supersedes the ./kernel subpath approach earlier in this branch. The base
entry '@canonical/task' is now node-free: types, the task monad, effect
constructors, primitives, combinators, the dry-run/testing interpreters,
undo collection (collectUndos), and TaskExecutionError. The node-touching
production and undo interpreters (run, runTask, executeEffect, runUndo) move
to '@canonical/task/node'.

TaskExecutionError is lifted to its own errors.ts (a plain Error subclass,
node-free) so dry-run and driveSync no longer reach the interpreter; undo
collection is split into undo.ts (node-free, base) from the executing
undo-interpreter.ts (node). The closure-walking test now pins the BASE
entry node-free. ./kernel and kernel.ts are removed.

BREAKING: consumers of run/runTask/executeEffect/runUndo import from
'@canonical/task/node'; test mocks that stub those symbols must target
'@canonical/task/node' (a base-module mock no longer intercepts them). All
in-repo consumers swept (cli/core, cli/pragma, cli/summon, summon/*):
import redirects and two mock-target repoints, base symbols unchanged.
