+++ packages/cli/core/src/lib/runGeneratorTask.ts
@@ -10,7 +10,8 @@
  * added without touching the callers.
  */
 
-import { type RunTaskOptions, runTask, type Task } from "@canonical/task";
+import { type Task } from "@canonical/task";
+import { type RunTaskOptions, runTask } from "@canonical/task/node";
 
 /** Options controlling how {@link runGeneratorTask} executes a task. */
 export interface RunGeneratorTaskOptions {
