+++ packages/cli/pragma/src/domains/doctor/operations/checks/checkMcpCommands.ts
@@ -2,7 +2,7 @@ import { existsSync } from "node:fs";
 import { delimiter, isAbsolute, join } from "node:path";
 import type { DetectedHarness } from "@canonical/harnesses";
 import { detectHarnesses, readMcpConfig } from "@canonical/harnesses";
-import { runTask } from "@canonical/task";
+import { runTask } from "@canonical/task/node";
 import type { CheckContext, CheckResult } from "../types.js";
 
 const NAME = "MCP commands";
