+++ packages/cli/pragma/src/domains/doctor/operations/checks/checkSkillsSymlinked.ts
@@ -1,7 +1,7 @@
 import { existsSync } from "node:fs";
 import type { DetectedHarness } from "@canonical/harnesses";
 import { detectHarnesses } from "@canonical/harnesses";
-import { runTask } from "@canonical/task";
+import { runTask } from "@canonical/task/node";
 import type { CheckContext, CheckResult } from "../types.js";
 
 /**
