+++ packages/runtime/task/README.md
@@ -12,17 +12,18 @@ bun add @canonical/task
 
 ## Entry points
 
-The package ships two entries:
+The package ships two entries, split by what they reach:
 
-- **`@canonical/task`** — everything: the task algebra plus the interpreters (`runTask`, dry-run, undo). The production interpreter executes filesystem, exec, and prompt effects, so this entry imports Node builtins (`node:fs/promises`, `node:path`, and `node:child_process` for `exec`) and is Node-only.
-- **`@canonical/task/kernel`** — the node-free algebra: core types, the task monad, effect constructors, primitives, and combinators. No Node builtin is reachable from this entry's import graph (pinned by a closure-walking test), so it bundles cleanly for browser and edge targets. Use it anywhere tasks are constructed, composed, or inspected but interpreted elsewhere.
+- **`@canonical/task`** — the node-free base: core types, the task monad, effect constructors, primitives, combinators, the dry-run/testing interpreters (`dryRun`, `dryRunWith`, `collectEffects`, …), undo collection (`collectUndos`), and `TaskExecutionError`. No Node builtin is reachable from this entry's import graph (pinned by a closure-walking test), so it bundles cleanly for browser and edge targets. Everything about *describing*, *composing*, and *previewing* work lives here.
+- **`@canonical/task/node`** — the node-touching interpreters: `runTask` / `run` / `executeEffect` (which execute filesystem, exec, and prompt effects, importing `node:fs/promises`, `node:path`, and `node:child_process`) and `runUndo`. Only actually *running* a task against the host needs this entry.
 
-Bundlers targeting the browser must import from `@canonical/task/kernel`. Module resolution happens before tree-shaking, so importing even a single pure function from the root entry pulls the interpreter's `node:` imports into the module graph and fails the build — regardless of what the bundle actually uses.
+The split matters because module resolution happens before tree-shaking: if the interpreters lived in the base entry, importing even a single pure function would pull `node:` imports into the module graph and fail any browser-platform build — regardless of what the bundle actually uses.
 
 ## Quick Start
 
 ```typescript
-import { gen, $, writeFile, readFile, info, runTask } from "@canonical/task";
+import { gen, $, writeFile, readFile, info } from "@canonical/task";
+import { runTask } from "@canonical/task/node";
 
 const setup = gen(function* () {
   const name = yield* $(readFile("name.txt"));
@@ -394,7 +395,7 @@ Tasks are inert data until an interpreter walks the structure and decides what t
 Executes effects against real I/O — filesystem, processes, prompts:
 
 ```typescript
-import { runTask } from "@canonical/task";
+import { runTask } from "@canonical/task/node";
 
 const value = await runTask(myTask);
 ```
@@ -427,7 +428,8 @@ await runTask(myTask, {
 When a task fails, `runTask` throws a `TaskExecutionError` that wraps the `TaskError`:
 
 ```typescript
-import { runTask, TaskExecutionError } from "@canonical/task";
+import { TaskExecutionError } from "@canonical/task";
+import { runTask } from "@canonical/task/node";
 
 try {
   await runTask(myTask);
@@ -501,7 +503,8 @@ getAffectedFiles(effects);
 Walks the task tree with mocked forward effects (like `dryRun`), collects the `undo` task attached to each effect, then executes them in reverse (LIFO) order. This enables `--undo` on any CLI command without storing state — the same task definition plus the same answers yields deterministic undo.
 
 ```typescript
-import { runUndo, collectUndos } from "@canonical/task";
+import { collectUndos } from "@canonical/task";
+import { runUndo } from "@canonical/task/node";
 
 // Forward run:
 await runTask(generator.generate(answers));
@@ -599,9 +602,10 @@ result.toNotWriteFile("output/secret.key");
 | **Error handling** | `retry`, `retryWithBackoff`, `orElse`, `optional`, `attempt` |
 | **Resources** | `bracket`, `ensure` |
 | **Utilities** | `tap`, `tapError`, `delay`, `timeout`, `fold`, `zip`, `zip3` |
-| **Production interpreter** | `runTask`, `run`, `executeEffect`, `TaskExecutionError`, `RunTaskOptions` |
+| **Production interpreter** (`@canonical/task/node`) | `runTask`, `run`, `executeEffect`, `RunTaskOptions` |
 | **Dry-run interpreter** | `dryRun`, `dryRunWith`, `mockEffect`, `collectEffects`, `countEffects`, `filterEffects`, `getFileWrites`, `getAffectedFiles`, `assertEffects`, `assertFileWrites`, `expectTask` |
-| **Undo interpreter** | `runUndo`, `collectUndos`, `UndoResult`, `UndoOptions` |
+| **Undo** | `collectUndos` (base), `UndoOptions` (base); `runUndo`, `UndoResult` (`@canonical/task/node`) |
+| **Errors** | `TaskExecutionError` (base — thrown by the interpreters, catchable anywhere) |
 | **Effect constructors** | `readFileEffect`, `writeFileEffect`, `appendFileEffect`, `copyFileEffect`, `copyDirectoryEffect`, `deleteFileEffect`, `deleteDirectoryEffect`, `makeDirEffect`, `existsEffect`, `symlinkEffect`, `globEffect`, `execEffect`, `promptEffect`, `logEffect`, `readContextEffect`, `writeContextEffect`, `parallelEffect`, `raceEffect` |
 | **Effect utilities** | `describeEffect`, `isWriteEffect`, `getAffectedPaths` |
