+++ packages/runtime/task/package.json
@@ -10,9 +10,9 @@
       "types": "./dist/types/index.d.ts",
       "import": "./dist/esm/index.js"
     },
-    "./kernel": {
-      "types": "./dist/types/kernel.d.ts",
-      "import": "./dist/esm/kernel.js"
+    "./node": {
+      "types": "./dist/types/node.d.ts",
+      "import": "./dist/esm/node.js"
     }
   },
   "files": [
