+++ packages/runtime/task/vitest.config.ts
@@ -8,7 +8,13 @@ export default defineConfig({
     coverage: {
       provider: "v8",
       include: ["src/**/*.ts"],
-      exclude: ["**/index.ts", "**/*.test.ts", "**/*.d.ts", "**/types.ts"],
+      exclude: [
+        "**/index.ts",
+        "**/node.ts",
+        "**/*.test.ts",
+        "**/*.d.ts",
+        "**/types.ts",
+      ],
       thresholds: {
         statements: 100,
         branches: 100,
