+++ packages/summon/core/src/stamp/runTaskWithStamp.ts
@@ -1,4 +1,5 @@
-import { type Effect, runTask, type Task } from "@canonical/task";
+import { type Effect, type Task } from "@canonical/task";
+import { runTask } from "@canonical/task/node";
 import applyStamp from "./applyStamp.js";
 import type { RunTaskWithStampOptions } from "./types.js";
